+++ src/main/java/com/cobber/fta/plugins/LogicalTypePhoneNumber.java
@@ -19,6 +19,7 @@ import java.util.Locale;
 import java.util.Map;
 
 import com.cobber.fta.AnalysisConfig;
+import com.cobber.fta.AnalyzerContext;
 import com.cobber.fta.FactsTypeBased;
 import com.cobber.fta.LogicalTypeInfinite;
 import com.cobber.fta.PluginDefinition;
@@ -102,8 +103,8 @@ public class LogicalTypePhoneNumber extends LogicalTypeInfinite  {
 	}
 
 	@Override
-	public String isValidSet(final String dataStreamName, final long matchCount, final long realSamples, String currentRegExp, final FactsTypeBased facts, final Map<String, Long> cardinality, final Map<String, Long> outliers, final Shapes shapes, AnalysisConfig analysisConfig) {
-		if (getHeaderConfidence(dataStreamName) == 0 && cardinality.size() <= 20 || getConfidence(matchCount, realSamples, dataStreamName) < getThreshold()/100.0)
+	public String isValidSet(final AnalyzerContext context, final long matchCount, final long realSamples, String currentRegExp, final FactsTypeBased facts, final Map<String, Long> cardinality, final Map<String, Long> outliers, final Shapes shapes, AnalysisConfig analysisConfig) {
+		if (getHeaderConfidence(context.getStreamName()) == 0 && cardinality.size() <= 20 || getConfidence(matchCount, realSamples, context.getStreamName()) < getThreshold()/100.0)
 			return REGEXP;
 		return null;
 	}