+++ src/test/java/com/cobber/fta/PluginPercent.java
@@ -76,7 +76,7 @@ public class PluginPercent extends LogicalTypeInfinite {
 	}
 
 	@Override
-	public String isValidSet(final String dataStreamName, final long matchCount, final long realSamples,
+	public String isValidSet(final AnalyzerContext context, final long matchCount, final long realSamples,
 			String currentRegExp, final FactsTypeBased facts, final Map<String, Long> cardinality, final Map<String, Long> outliers, Shapes shapes, AnalysisConfig analysisConfig) {
 		return (double)matchCount/realSamples >= getThreshold()/100.0 ? null : ".+";
 	}